fix: consolidar sync/status + sync/mobne em api/sync.js (limite 12 funções Vercel Hobby)
