+++ frontend/src/services/mobneSync.js
@@ -2,7 +2,7 @@ import api from './api.js';
 
 export const mobneSync = {
   triggerSync: (entity = 'produtos', page = 1) =>
-    api.post('/sync/mobne', { entity, page, pageSize: 100 }),
+    api.post('/sync', { entity, page, pageSize: 100 }),
 
-  getStatus: () => api.get('/sync/status'),
+  getStatus: () => api.get('/sync'),
 };
